Add resource object
Resource object is a real work-horse of this redfish client. In its
most trivial incarnation, resource is a simple wrapper around hash
that allows us to access hash values using dot-notation.

Second thing that this object does is on-demand loading of
sub-resources. This means that any wrapped hash values that are also
hashes get automatically wrapped into resource and arrays of hashes
are converted into array of resources.

Resources also know how to handle OpenData ids and can, with the help
of connector object, fetch subresource data from remote service and
cache it locally.