Make sure loaded resource has @odata.id field
OData standard requires that all resources returned from OData
compliant services contain @odata.id field.

In practice, some of the services fail to do so, but since we are
already loading the resource based on the @odata.id values, adding
this to constructed resource is trivial and does not break any
existing code.