Allow resources to be fetched directly by OData id
Up until now, if we wanted to interact with a specific resource, we
had to navigate to it from root. Storing resource's id was of no use,
since we were not able to retrieve it without traversing and filtering
the resource tree.

This commit adds a method to the root resource that can be used to
directly fetch the resource by its OData id. This makes it possible to
save the resource's id and later use it to reconstruct the resource.

One use case for this functionality is changing the state of the
location LED on the selected server. Up until now, in order to turn
the LED on, we were force to iterate the systems toplevel collection
and match the id. Using new find method, iteration is not needed and
we can fetch the server directly.