Implement basic HTTP authentication support
Up until now, Redfish client only supported session authentication,
since it is mandatory as per Redfish standard [1]. This means that
client can interact with any Redfish-compliant service.

Unfortunately, services in the wild often only implement part of the
standard and in case of authentication, some of the services only
offer basic authentication, which makes our client useless for such
cases.

To bypass this, we added basic authentication support. By default, we
try to use session authentication if possible, since it is more secure
of the two, and only use basic authentication if session cannot be
created.

[1] http://redfish.dmtf.org/schemas/DSP0266_1.4.0.html#authentication