Properly isolate test cases
Previously, test cases reused some common components that should
remain unchanged during the testing. But since this immutability is
not enforced, we were relying on developers to do the RightThing(TM).

This commit completely decouples test cases by providing fresh copies
of the components that the test case is using.