Get rid of GET call to Sessions endpoint
Up until now, we created new session authentication token by
retrieving the Sessions resource and posting credentials through this
resource to obtain the authentication token. But this is not a viable
strategy, since standard [1] mandates that Sessions endpoint is only
available when authenticated, exception being the POST requests.

This commit fixes the login functionality by bypassing the resource
fetching mechanism and issuing a POST request directly to the Sessions
endpoint. Tests were also updated to follow the standard better.

[1] http://redfish.dmtf.org/schemas/DSP0266_1.4.0.html#http-header-security