Move JSON serialization into connector
Up until now, connector was not aware of the content type of the
payload that POST and PATCH requests were sending. This meant that all
connector consumers were expected to properly encode the data.

Since Redfish standard mandates that any payload is JSON, it makes
little sense to keep forcing consumers to encode the data that we know
that must be serialized to JSON in the end, so we moved this
serialization into the connector class.

This also resolves the issue of not being able to set the Content-Type
header correctly since connector did not expose this kind of
functionality. Now, content type is set to application/json if the
payload is present.