Allow client to follow redirects
Up until now, client did not follow redirects automatically, since
none of the Redfish services that we connected to returned a redirect
status.

Testing the client on the live Redfish service revealed that we can
get redirect response that need to be handled properly. Fortunately,
Excon already provides redirect middleware that just needs to be
activated. And this is what we did in this commit.