Use the same version of rubocop in CI and locally
Debugging package issues can be a bit frustrating if the CI machinery
is running a different versions of helper tools, since configuration
can and do change as the tools mature.

This commit gets rubocop versions on the CI and in the local
development machine back in sync by pinning the version in gemfile and
explicitly specifying the channel to use on CodeClimate.