Merge connectors into single class
Up until now, we had a two separate classes of connectors: one that
cached GET responses and one that did not. In order to make it easier
to extend the connector, we merged both of those classes into a single
class.

In order to keep things loosely coupled, we use dependency injection
to set the caching backend.

Part of the unification was also unit test refactor, since it was
getting near impossible to add new tests without studying the whole
spec file and getting acquainted with complete test mock setup.

New tests use webmock's rspec matchers that make it possible to define
the test cases in really readable way.

We also removed all shared state from the tests. Test cases are now a
bit bulkier, but far easier to comprehend and extend.