Abstract away connector response
Up until now connector was returning the response object of the
underlying HTTP library, which was not optimal, since we were using
somewhat raw types far beyond the connector class, making our library
fragile and tightly coupled.

In this commit, we introduced new response class that is local to our
connector. This object also abstract away some of the underlying HTTP
library details and provides an uniform API for the higher layers of
Redfish library.