Move login functionality from root to connector
We were treating the connector as a dumb HTTP interface to the API
until now, which was great, since connector knew nothing about the API
semantics. But this approach also made some things almost impossible
to properly fix.

Main use case that this change addresses is login session expiration.
Since connector knew nothing about the sessions and login in general,
library consumers were forced to handle session expiration on their
own. Needless to say, this was fiddly and error-prone.