Implement automatic login retry
If the user session that client created expired during operation,
previous versions of client simply returned the error response to the
caller, which was not optimal since it required a lot of manual
handling around the API calls.

This commit introduces the automatic login retry that should remove
the need to handle errors in wast majority of use cases.