Add RedfishClient::Resource#refresh method
This method is a replacement for removed reset method that was working
at the wrong level ob abstraction.

Now anytime we want to be sure that resource data is in sync with the
one on the API endpoint, we simply call refresh method and resource
will take appropriate steps to invalidate cache and fetch fresh data.