Refactor internal request execution
In order to simplify the internal structure of the library, we created
a more direct path for request execution. This means that the only two
methods that have any kind of logic inside them are Connector#request
and Resource#request. All other methods use those two and are offered
as a convenience to the library end-users.