Add a use_session parameter to force basic auth usage
It's important to clean up sessions because various Redfish
implementations have a limit on the number of sessions. Forgetting to
clean up can have a DoS effect.

Further testing showed that, at least in some implementations, there's a
significant overhead on setting up a session for a single request while
for multiple requests it's unclear if it provides benefit.

                                               user       system     total       real
curl on root without login:                    0.000457   0.000111   0.063841 (  0.719540)
curl on root with basic auth:                  0.000389   0.000095   0.048539 (  0.807732)
curl on root with session auth:                0.000479   0.000116   0.186354 (  2.672784)
curl on multiple endpoints with session auth:  0.000859   0.000209   0.480265 ( 10.841015)
curl on multiple endpoints with basic auth:    0.000726   0.000177   0.367814 (  8.859796)

                                               user       system     total       real
curl on root without login:                    0.000427   0.000096   0.057136 (  0.687104)
curl on root with basic auth:                  0.000467   0.000104   0.048592 (  0.862687)
curl on root with session auth:                0.000421   0.000094   0.178445 (  2.568848)
curl on multiple endpoints with session auth:  0.001025   0.000229   0.576333 (  8.948681)
curl on multiple endpoints with basic auth:    0.000000   0.001359   0.450287 ( 10.347542)

                                               user       system      total      real
curl on root without login:                    0.000448   0.000073   0.054231 (  0.678578)
curl on root with basic auth:                  0.000406   0.000065   0.049842 (  0.867248)
curl on root with session auth:                0.000477   0.000078   0.178306 (  2.403250)
curl on multiple endpoints with session auth:  0.000971   0.000158   0.518806 (  9.558119)
curl on multiple endpoints with basic auth:    0.001447   0.000000   0.405758 ( 10.668350)

This implements a use_session parameter to force basic auth usage,
giving control over this.